+++ docs/release-notes.rst
@@ -1,11 +1,28 @@
 ==================================
 Release Notes
 ==================================
+*************
+Version 0.7.0
+*************
+Release Date 24/07/2026
+
+Major Changes
+################
+- New major version! This should have been done in the last update when python 3.8 to 3.11 was dropped. Doing it here now.
+
+New Features
+############
+- Font size for buttons increased.
+- Image scene now shows "TEST PASS" or "TEST FAIL" on sequence end.
+- Added --user-id flag to be able to log which user ran a test. Defaults to 'None' if not provided.
+- Added aliases for "quit" and "open" commands in fxconfig.
+
+
 
 *************
 Version 0.6.5
 *************
-Release Date XX/XX/XXXX
+Release Date 29/06/2026
 
 Major Changes
 ################
@@ -26,9 +43,11 @@ Improvements
 - Created a new FixateError base class for all exceptions raised by fixate to use. It inherits from Exception instead of BaseExcepetion to improve error handling.
 - DSO Driver function 'waveform_values' now returns a single channels x and y data as two separate lists, without re-acquiring the signal. This function should
   now be called after performing signal acquisition.
-- Invert channel and vtime functions implemented in the DSO driver.
+- UI has been revamped a little. User input methods will no longer return ("Resp", value) but will return just the value - see Issue #177. To use these, from fixate import user_info etc.
+- Fixed a bug where falsy values such as 0 or 0.0 were not accepted as valid user input - see Issue #213
+- Invert channel and vtime funcitons implemented in the DSO driver.
 - Fxconfig fixes for cmd2 4.0.0
-- Added aliases for "quit" and "open" commands in fxconfig.
+- Jig switching fix to force sending the reset signal regardless of presumed jig state
 
 *************
 Version 0.6.4